Move VerifyPage into sub-folder
